account/account/requirements.txt
Django==4.0.4
Pillow==9.1.0
PyJWT==2.3.0
loguru==0.6.0
mysqlclient==2.1.0
pytest==7.1.1
